+++ Assets/Atlases/Particles.json
@@ -74,8 +74,8 @@
     },
     "BigDisperse": {
       "frame": {
-        "x": 0,
-        "y": 512,
+        "x": 512,
+        "y": 0,
         "w": 768,
         "h": 64
       },
@@ -99,79 +99,7 @@
     "FireSmoke": {
       "frame": {
         "x": 512,
-        "y": 0,
-        "w": 768,
-        "h": 64
-      },
-      "rotated": false,
-      "trimmed": false,
-      "spriteSourceSize": {
-        "x": 0,
-        "y": 0,
-        "w": 768,
-        "h": 64
-      },
-      "sourceSize": {
-        "w": 768,
-        "h": 64
-      },
-      "pivot": {
-        "x": 0.5,
-        "y": 0.5
-      }
-    },
-    "GroundPuff": {
-      "frame": {
-        "x": 512,
         "y": 64,
-        "w": 768,
-        "h": 64
-      },
-      "rotated": false,
-      "trimmed": false,
-      "spriteSourceSize": {
-        "x": 0,
-        "y": 0,
-        "w": 768,
-        "h": 64
-      },
-      "sourceSize": {
-        "w": 768,
-        "h": 64
-      },
-      "pivot": {
-        "x": 0.5,
-        "y": 0.5
-      }
-    },
-    "Ring": {
-      "frame": {
-        "x": 512,
-        "y": 128,
-        "w": 768,
-        "h": 64
-      },
-      "rotated": false,
-      "trimmed": false,
-      "spriteSourceSize": {
-        "x": 0,
-        "y": 0,
-        "w": 768,
-        "h": 64
-      },
-      "sourceSize": {
-        "w": 768,
-        "h": 64
-      },
-      "pivot": {
-        "x": 0.5,
-        "y": 0.5
-      }
-    },
-    "Simple": {
-      "frame": {
-        "x": 512,
-        "y": 192,
         "w": 768,
         "h": 64
       },
@@ -195,6 +123,54 @@
     "WideFloor": {
       "frame": {
         "x": 512,
+        "y": 128,
+        "w": 768,
+        "h": 64
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 768,
+        "h": 64
+      },
+      "sourceSize": {
+        "w": 768,
+        "h": 64
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "Ring": {
+      "frame": {
+        "x": 512,
+        "y": 192,
+        "w": 768,
+        "h": 64
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 768,
+        "h": 64
+      },
+      "sourceSize": {
+        "w": 768,
+        "h": 64
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "Simple": {
+      "frame": {
+        "x": 512,
         "y": 256,
         "w": 768,
         "h": 64
@@ -216,11 +192,11 @@
         "y": 0.5
       }
     },
-    "GroundPlume": {
+    "GroundPuff": {
       "frame": {
         "x": 512,
         "y": 320,
-        "w": 704,
+        "w": 768,
         "h": 64
       },
       "rotated": false,
@@ -228,11 +204,11 @@
       "spriteSourceSize": {
         "x": 0,
         "y": 0,
-        "w": 704,
+        "w": 768,
         "h": 64
       },
       "sourceSize": {
-        "w": 704,
+        "w": 768,
         "h": 64
       },
       "pivot": {
@@ -240,7 +216,7 @@
         "y": 0.5
       }
     },
-    "GroundRing": {
+    "SmallPuff": {
       "frame": {
         "x": 512,
         "y": 384,
@@ -264,7 +240,7 @@
         "y": 0.5
       }
     },
-    "SmallPuff": {
+    "GroundRing": {
       "frame": {
         "x": 512,
         "y": 448,
@@ -291,6 +267,30 @@
     "TallGroundPlume": {
       "frame": {
         "x": 0,
+        "y": 512,
+        "w": 704,
+        "h": 64
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 704,
+        "h": 64
+      },
+      "sourceSize": {
+        "w": 704,
+        "h": 64
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "GroundPlume": {
+      "frame": {
+        "x": 0,
         "y": 576,
         "w": 704,
         "h": 64
@@ -306,6 +306,174 @@
       "sourceSize": {
         "w": 704,
         "h": 64
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "DashedCircle": {
+      "frame": {
+        "x": 1216,
+        "y": 384,
+        "w": 64,
+        "h": 64
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 64,
+        "h": 64
+      },
+      "sourceSize": {
+        "w": 64,
+        "h": 64
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "VortexRing": {
+      "frame": {
+        "x": 1216,
+        "y": 448,
+        "w": 64,
+        "h": 64
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 64,
+        "h": 64
+      },
+      "sourceSize": {
+        "w": 64,
+        "h": 64
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "InternalFadeCircle": {
+      "frame": {
+        "x": 704,
+        "y": 512,
+        "w": 64,
+        "h": 64
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 64,
+        "h": 64
+      },
+      "sourceSize": {
+        "w": 64,
+        "h": 64
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "SharpTear": {
+      "frame": {
+        "x": 704,
+        "y": 576,
+        "w": 36,
+        "h": 36
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 36,
+        "h": 36
+      },
+      "sourceSize": {
+        "w": 36,
+        "h": 36
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "SpikyCircle": {
+      "frame": {
+        "x": 740,
+        "y": 576,
+        "w": 32,
+        "h": 32
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 32,
+        "h": 32
+      },
+      "sourceSize": {
+        "w": 32,
+        "h": 32
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "StarX": {
+      "frame": {
+        "x": 740,
+        "y": 608,
+        "w": 32,
+        "h": 32
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 32,
+        "h": 32
+      },
+      "sourceSize": {
+        "w": 32,
+        "h": 32
+      },
+      "pivot": {
+        "x": 0.5,
+        "y": 0.5
+      }
+    },
+    "GlowStar": {
+      "frame": {
+        "x": 768,
+        "y": 512,
+        "w": 32,
+        "h": 32
+      },
+      "rotated": false,
+      "trimmed": false,
+      "spriteSourceSize": {
+        "x": 0,
+        "y": 0,
+        "w": 32,
+        "h": 32
+      },
+      "sourceSize": {
+        "w": 32,
+        "h": 32
       },
       "pivot": {
         "x": 0.5,
