+++ Content/Bosses/KingSlime/KingSlime.cs
@@ -5,6 +5,7 @@ using AcidicBosses.Common;
 using AcidicBosses.Common.Configs;
 using AcidicBosses.Common.Effects;
 using AcidicBosses.Content.Particles;
+using AcidicBosses.Content.Particles.Animated;
 using AcidicBosses.Helpers;
 using Luminance.Common.Utilities;
 using Luminance.Core.Graphics;
