+++ Content/Bosses/Skeletron/SkeletronHead.cs
@@ -3,6 +3,7 @@ using System.IO;
 using AcidicBosses.Common.Configs;
 using AcidicBosses.Common.Effects;
 using AcidicBosses.Content.Particles;
+using AcidicBosses.Content.Particles.Animated;
 using AcidicBosses.Content.ProjectileBases;
 using AcidicBosses.Content.Projectiles;
 using AcidicBosses.Core.StateManagement;
