+++ Content/Bosses/Twins/TwinsController.cs
@@ -3,15 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using AcidicBosses.Common.Textures;
-using AcidicBosses.Content.Bosses.Twins.Projectiles;
-using AcidicBosses.Content.Bosses.WoF.Projectiles;
-using AcidicBosses.Content.Particles;
-using AcidicBosses.Content.ProjectileBases;
-using AcidicBosses.Content.Projectiles;
 using AcidicBosses.Core.StateManagement;
-using AcidicBosses.Helpers;
-using AcidicBosses.Helpers.NpcHelpers;
-using Luminance.Common.Utilities;
 using Luminance.Common.VerletIntergration;
 using Luminance.Core.Graphics;
 using Microsoft.Xna.Framework;
@@ -19,16 +11,14 @@ using Microsoft.Xna.Framework.Graphics;
 using ReLogic.Content;
 using ReLogic.Graphics;
 using Terraria;
-using Terraria.Audio;
 using Terraria.GameContent;
 using Terraria.ID;
 using Terraria.ModLoader;
-using Animation = AcidicBosses.Core.Animation.Animation;
 using Vector2 = Microsoft.Xna.Framework.Vector2;
 
 namespace AcidicBosses.Content.Bosses.Twins;
 
-public class TwinsController : AcidicNPC
+public partial class TwinsController : AcidicNPC
 {
     public override string Texture => TextureRegistry.InvisPath;
 
@@ -85,7 +75,7 @@ public class TwinsController : AcidicNPC
 
         phaseTracker = new PhaseTracker([
             PhaseUntransformed,
-            PhaseTranformation,
+            PhaseTransformation,
             PhaseTransformed1
         ]);
 
@@ -156,751 +146,6 @@ public class TwinsController : AcidicNPC
     }
 
     #endregion
-    
-    #region Phases
-
-    private PhaseState PhaseUntransformed => new PhaseState(Phase_Untransformed, EnterPhaseUntransformed);
-    private PhaseState PhaseTranformation => new(Phase_Transformation, EnterPhaseTransformation);
-    private PhaseState PhaseTransformed1 => new PhaseState(Phase_Transformed1, EnterPhaseTransformed1);
-    
-    private void EnterPhaseUntransformed()
-    {
-        var hover = new AttackState(() => Attack_Hover(90, 20, 0.25f), 30);
-        var doubleDash = new AttackState(() => Attack_DoubleDash(10, 30, 15, 30), 60);
-        var longCrossDash = new AttackState(() => Attack_CrossDash(20, 60, 30), 0);
-        var plusDash = new AttackState(() => Attack_PlusDash(20, 60, 25), 0);
-        var crossDash = new AttackState(() => Attack_CrossDash(20, 60, 25), 0);
-        var alternatingDashes = new AttackState(() => Attack_AlternatingDashes(60 * 4, 10, 30, 15, 20), 0);
-        var fastAlternatingDashes = new AttackState(() => Attack_AlternatingFastDashes(60 * 3, 10, 60, 5, 10), 0);
-        var recenter = new AttackState(Attack_Recenter, 0);
-        
-        attackManager.SetAttackPattern([
-            hover,
-            doubleDash, doubleDash,
-            alternatingDashes,
-            fastAlternatingDashes,
-            recenter,
-            hover,
-            doubleDash, doubleDash, doubleDash,
-            hover,
-            longCrossDash, plusDash, crossDash, plusDash, crossDash,
-            recenter,
-            hover,
-            alternatingDashes
-        ]);
-    }
-    
-    private void Phase_Untransformed()
-    {
-        if (attackManager.InWindDown)
-        {
-            if (AverageLifePercent <= 1f) // TODO Temp
-            {
-                phaseTracker.NextPhase();
-                attackManager.Reset();
-                return;
-            }
-            
-            Attack_Hover(20f, 0.25f);
-            return;
-        }
-        
-        attackManager.RunAttackPattern();
-    }
-
-    private Animation transformation;
-    private void EnterPhaseTransformation()
-    {
-        transformation = new Animation();
-        
-        // Stop Moving
-        transformation.AddConstantEvent((progress, frame) =>
-        {
-            Spazmatism.Npc.SimpleFlyMovement(Vector2.Zero, 0.2f);
-            Retinazer.Npc.SimpleFlyMovement(Vector2.Zero, 0.2f);
-        });
-        
-        // Start rumble
-        transformation.AddInstantEvent(0, () =>
-        {
-            ScreenShakeSystem.SetUniversalRumble(0.5f, shakeStrengthDissipationIncrement: 0f);
-        });
-
-        // Make Smoke
-        transformation.AddTimedEvent(0, 120, (progress, frame) =>
-        {
-            DoBothTwins(twin =>
-            {
-                var npc = twin.Npc;
-                if (frame % 30 == 0)
-                {
-                    SoundEngine.PlaySound(SoundID.Item13, npc.Center);
-                }
-
-                var smokeDusts = MathHelper.Lerp(1f, 4f, progress);
-                for (var i = 0; i < smokeDusts; i++)
-                {
-                    var speed = Main.rand.NextVector2Circular(5, 5);
-                    var pos = Main.rand.NextVector2Circular(npc.width / 2f, npc.height / 2f) + npc.Center;
-                    Dust.NewDust(pos, 0, 0, DustID.Smoke, speed.X, speed.Y, Scale: 1.5f);
-                }
-                if (Main.rand.NextBool(smokeDusts % 1f))
-                {
-                    var speed = Main.rand.NextVector2Circular(5, 5);
-                    var pos = Main.rand.NextVector2Circular(npc.width / 2f, npc.height / 2f) + npc.Center;
-                    Dust.NewDust(pos, 0, 0, DustID.Smoke, speed.X, speed.Y, Scale: 1.5f);
-                }
-                
-                var smokeParticles = MathHelper.Lerp(0f, 1f, progress);
-                for (var i = 0; i < smokeParticles; i++)
-                {
-                    var speed = Main.rand.NextVector2Circular(2.5f, 2.5f);
-                    var pos = Main.rand.NextVector2Circular(npc.width / 2f, npc.height / 2f) + npc.Center;
-                    var rot = Main.rand.NextFloatDirection();
-                    var angularVel = Main.rand.NextFloat(0f, MathHelper.Pi / 16f);
-                    var puff = new SmallPuffParticle(pos, speed, rot, Color.WhiteSmoke, 30);
-                    puff.Opacity = 0.5f;
-                    puff.AngularVelocity = angularVel;
-                    puff.Spawn();
-                }
-                if (Main.rand.NextBool(smokeParticles % 1f))
-                {
-                    var speed = Main.rand.NextVector2Circular(2.5f, 2.5f);
-                    var pos = Main.rand.NextVector2Circular(npc.width / 2f, npc.height / 2f) + npc.Center;
-                    var rot = Main.rand.NextFloatDirection();
-                    var angularVel = Main.rand.NextFloat(0f, MathHelper.Pi / 16f);
-                    var puff = new SmallPuffParticle(pos, speed, rot, Color.WhiteSmoke, 30);
-                    puff.Opacity = 0.5f;
-                    puff.AngularVelocity = angularVel;
-                    puff.Spawn();
-                }
-            });
-        });
-        
-        // Transform twins
-        transformation.AddInstantEvent(120, () =>
-        {
-            ScreenShakeSystem.SetUniversalRumble(0f, shakeStrengthDissipationIncrement: 0.2f);
-            ScreenShakeSystem.StartShake(2f);
-            
-            DoBothTwins(twin =>
-            {
-                var npc = twin.Npc;
-            
-                SoundEngine.PlaySound(SoundID.NPCHit1, npc.Center);
-                SoundEngine.PlaySound(SoundID.Roar, npc.Center);
-
-                var burst = new BigSmokeDisperseParticle(npc.Bottom, Vector2.Zero, 0f, Color.WhiteSmoke, 60);
-                burst.IgnoreLighting = true;
-                burst.Scale *= 2.25f;
-                burst.Spawn();
-            
-                for (var i = 0; i < 2; i++)
-                {
-                    Gore.NewGore(NPC.GetSource_FromAI(), npc.Center, Main.rand.NextVector2Square(-30, 31) * 0.2f, 144);
-                    Gore.NewGore(NPC.GetSource_FromAI(), npc.Center, Main.rand.NextVector2Square(-30, 31) * 0.2f, 7);
-                    Gore.NewGore(NPC.GetSource_FromAI(), npc.Center, Main.rand.NextVector2Square(-30, 31) * 0.2f, 6);
-                }
-                for (var i = 0; i < 20; i++)
-                {
-                    Dust.NewDust(npc.position, npc.width, npc.height, DustID.Blood, Main.rand.Next(-30, 31) * 0.2f, Main.rand.Next(-30, 31) * 0.2f);
-                }
-                
-                var dustId = twin is Spazmatism ? DustID.CursedTorch : DustID.CrimsonTorch;
-                for (var i = 0; i < 20; i++)
-                {
-                    var speed = Main.rand.NextVector2CircularEdge(15f, 15f);
-                    Dust.NewDust(npc.Center, 0, 0, dustId, speed.X, speed.Y, Scale: 2f);
-                }
-            
-                npc.HitSound = SoundID.NPCHit4;
-                twin.MechForm = true;
-            });
-        });
-    }
-
-    private void Phase_Transformation()
-    {
-        var done = transformation.RunAnimation();
-        if (done)
-        {
-            phaseTracker.NextPhase();
-            attackManager.Reset();
-        }
-    }
-    
-    private void EnterPhaseTransformed1()
-    {
-        var hover = new AttackState(() => Attack_Hover(45, 30, 0.355f), 30);
-        var doubleDash = new AttackState(() => Attack_DoubleDash(15, 30, 15, 30), 60);
-        var recenter = new AttackState(Attack_Recenter, 0);
-        var test = new AttackState(Attack_Test, 30);
-        var sweep = new AttackState(Attack_SweepingLaser, 30);
-        
-        attackManager.SetAttackPattern([
-            hover,
-            test, sweep
-        ]);
-    }
-    
-    private void Phase_Transformed1()
-    {
-        Spazmatism.MechForm = true;
-        Retinazer.MechForm = true;
-        
-        if (attackManager.InWindDown)
-        {
-            var target = Main.player[NPC.target];
-            Attack_Hover(30f, 0.3f);
-            return;
-        }
-        
-        attackManager.RunAttackPattern();
-    }
-    
-    #endregion
-    
-    #region Attacks
-    
-    private void Attack_Hover(Twin twin, float speed, float accel)
-    {
-        var target = Main.player[NPC.target];
-        if (twin == Spazmatism)
-        {
-            var spazOffset = new Vector2(-200, 0);
-            if (Spazmatism.Npc.Center.X > target.Center.X) spazOffset = new Vector2(200, 0);
-            FlyTo(Spazmatism.Npc, target.Center + spazOffset, speed, accel);
-            Spazmatism.LookTowards(target.Center, 0.05f);
-        }
-        else
-        {
-            var retOffset = new Vector2(0, -250);
-            if (Retinazer.Npc.Center.Y > target.Center.Y) retOffset = new Vector2(0, 250);
-            FlyTo(Retinazer.Npc, target.Center + retOffset, speed, accel);
-            Retinazer.LookTowards(target.Center, 0.05f);
-        }
-    }
-
-    private void Attack_Hover(float speed, float accel)
-    {
-        Attack_Hover(Spazmatism, speed, accel);
-        Attack_Hover(Retinazer, speed, accel);
-    }
-    
-    private bool Attack_Hover(int hoverTime, float speed, float acceleration)
-    {
-        attackManager.CountUp = true;
-
-        Attack_Hover(speed, acceleration);
-
-        if (attackManager.AiTimer <= hoverTime) return false;
-
-        attackManager.CountUp = false;
-        return true;
-    }
-    
-    private void Attack_Pull(Twin leadTwin)
-    {
-        switch (leadTwin.Npc.type)
-        {
-            case NPCID.Spazmatism:
-                SimulateTether(1, 10);
-                break;
-            case NPCID.Retinazer:
-                SimulateTether(10, 1);
-                break;
-            default:
-                throw new Exception($"{leadTwin.Npc.whoAmI} is somehow not a twin?");
-        }
-    }
-
-    private bool Attack_Recenter()
-    {
-        var target = Main.player[NPC.target];
-        
-        var spazOffset = new Vector2(-400, 0);
-        if (Spazmatism.Npc.Center.X > target.Center.X) spazOffset = new Vector2(400, 0);
-        Teleport(Spazmatism, target.Center + spazOffset, 20);
-        Spazmatism.LookTowards(target.Center, 0.05f);
-        
-        var retOffset = new Vector2(0, -450);
-        // if (Retinazer.Npc.Center.Y > target.Center.Y) retOffset = new Vector2(0, 450);
-        Teleport(Retinazer, target.Center + retOffset, 20);
-        Retinazer.LookTowards(target.Center, 0.05f);
-
-        return true;
-    }
-
-    private DashState Attack_Dash(Twin twin, int dashLength, float speed, int trackTime, int dashAtTime, float distance)
-    {
-        var options = new DashOptions
-        {
-            MinimumDistance = distance,
-            DashSpeed = speed,
-            DashLength = dashLength,
-            LookOffset = MathHelper.PiOver2,
-            TrackTime = trackTime,
-            DashAtTime = dashAtTime
-        };
-        
-        var target = Main.player[NPC.target];
-
-        var dashState = Dash(twin, twin.AttackManager, target.Center, options);
-
-        return dashState;
-    }
-    
-    private DashState Attack_TrackingDash(Twin twin, int dashLength, float speed, int trackTime, int dashAtTime, float distance)
-    {
-        var options = new DashOptions
-        {
-            MinimumDistance = distance,
-            DashSpeed = speed,
-            DashLength = dashLength,
-            LookOffset = MathHelper.PiOver2,
-            TrackTime = trackTime,
-            DashAtTime = dashAtTime
-        };
-        
-        var npc = twin.Npc;
-        var target = Main.player[NPC.target];
-        
-        var trackedPos = target.Center;
-        trackedPos += target.velocity * trackTime; // Lead player pos
-        var travelTime = npc.Distance(trackedPos) / speed;
-        trackedPos += target.velocity * travelTime; // Account for travel time
-
-        var dashState = Dash(twin, twin.AttackManager, trackedPos, options);
-        
-        return dashState;
-    }
-    
-    private bool Attack_DoubleDash(int dashLength, float speed, int trackTime, int dashAtTime)
-    {
-        attackManager.CountUp = true;
-
-        ref var spazDone = ref NPC.localAI[0];
-        ref var retDone = ref NPC.localAI[1];
-        if (attackManager.AiTimer == 0)
-        {
-            Spazmatism.AttackManager.AiTimer = 0;
-            Retinazer.AttackManager.AiTimer = 0;
-            spazDone = 0;
-            retDone = 0;
-        }
-        
-        // Spaz has a normal dash while Ret tracks the player and dashes later
-        if (spazDone == 0)
-        {
-            var spazDashState = Attack_Dash(Spazmatism, dashLength, speed, trackTime, dashAtTime, 300);
-            if (spazDashState == DashState.Done) spazDone = 1;
-        }
-
-        if (retDone == 0)
-        {
-            var retDashState = Attack_TrackingDash(Retinazer, dashLength, speed, trackTime * 2, (int) (dashAtTime * 2f), 300);
-            if (retDashState == DashState.Done) retDone = 1;
-        }
-        
-        if (spazDone != 0) Attack_Hover(Spazmatism, 20, 0.25f);
-        if (retDone != 0) Attack_Hover(Retinazer, 20, 0.25f);
-
-        if (spazDone == 0 || retDone == 0) return false;
-        
-        spazDone = 0;
-        retDone = 0;
-        attackManager.CountUp = false;
-        return true;
-
-    }
-
-    private bool Attack_PlusDash(int dashLength, float speed, int dashAtTime)
-    {
-        var dashOptions = new DashOptions
-        {
-            DashLength = dashLength,
-            DashSpeed = speed,
-            LookOffset = MathHelper.PiOver2,
-            DashAtTime = dashAtTime,
-            MinimumDistance = 0,
-            DontReposition = true,
-            TrackTime = 0
-        };
-
-        var target = Main.player[NPC.target];
-
-        if (attackManager.AiTimer == 0)
-        {
-            var spazPos = target.Center + new Vector2(350, 0);
-            if (Spazmatism.Npc.Center.X < target.Center.X) spazPos = target.Center - new Vector2(350, 0);
-            var retPos = target.Center + new Vector2(0, 350);
-            if (Retinazer.Npc.Center.Y > target.Center.Y) retPos = target.Center - new Vector2(0, 350);
-            
-            Spazmatism.LookTowards(spazPos, 1f);
-            Retinazer.LookTowards(retPos, 1f);
-            
-            Teleport(Spazmatism, spazPos, 0);
-            Teleport(Retinazer, retPos, 0);
-            
-            Spazmatism.LookTowards(target.Center, 1f);
-            Retinazer.LookTowards(target.Center, 1f);
-            
-            Spazmatism.Npc.velocity = Vector2.Zero;
-            Retinazer.Npc.velocity = Vector2.Zero;
-        }
-        
-        // Dashes are perfectly synced together because there's no repositioning
-        var dashState = Dash(Spazmatism, attackManager, target.Center, dashOptions);
-        Dash(Retinazer, attackManager, target.Center, dashOptions);
-
-        return dashState == DashState.Done;
-    }
-    
-    private bool Attack_CrossDash(int dashLength, float speed, int dashAtTime)
-    {
-        var dashOptions = new DashOptions
-        {
-            DashLength = dashLength,
-            DashSpeed = speed,
-            LookOffset = MathHelper.PiOver2,
-            DashAtTime = dashAtTime,
-            MinimumDistance = 0,
-            DontReposition = true,
-            TrackTime = 0
-        };
-
-        var target = Main.player[NPC.target];
-
-        if (attackManager.AiTimer == 0)
-        {
-            var spazPos = target.Center + new Vector2(300, -300);
-            if (Spazmatism.Npc.Center.Y < target.Center.Y) spazPos = target.Center - new Vector2(300, -300);
-            var retPos = target.Center + new Vector2(-300, -300);
-            if (Retinazer.Npc.Center.Y < target.Center.Y) retPos = target.Center - new Vector2(-300, -300);
-            
-            Spazmatism.LookTowards(spazPos, 1f);
-            Retinazer.LookTowards(retPos, 1f);
-            
-            Teleport(Spazmatism, spazPos, 0);
-            Teleport(Retinazer, retPos, 0);
-            
-            Spazmatism.LookTowards(target.Center, 1f);
-            Retinazer.LookTowards(target.Center, 1f);
-            
-            Spazmatism.Npc.velocity = Vector2.Zero;
-            Retinazer.Npc.velocity = Vector2.Zero;
-        }
-        
-        // Dashes are perfectly synced together because there's no repositioning
-        var dashState = Dash(Spazmatism, attackManager, target.Center, dashOptions);
-        Dash(Retinazer, attackManager, target.Center, dashOptions);
-        
-        return dashState == DashState.Done;
-    }
-
-    private bool Attack_AlternatingDashes(int length, int dashLength, float speed, int trackTime, int dashAtTime)
-    {
-        attackManager.CountUp = true;
-        
-        ref var turn = ref NPC.localAI[0];
-        Twin dashingTwin = turn == 0 ? Spazmatism : Retinazer;
-        Twin notDashingTwin = turn == 0 ? Retinazer : Spazmatism;
-
-        Attack_Hover(notDashingTwin, 20f, 0.2f);
-        
-        var target = Main.player[NPC.target];
-        var dashOptions = new DashOptions
-        {
-            DashLength = dashLength,
-            DashSpeed = speed,
-            LookOffset = MathHelper.PiOver2,
-            DashAtTime = dashAtTime,
-            MinimumDistance = 250,
-            TrackTime = trackTime
-        };
-
-        var dashState = Dash(dashingTwin, dashingTwin.AttackManager, target.Center, dashOptions);
-
-        if (dashState == DashState.Done)
-        {
-            turn = ((int) turn + 1) % 2;
-            dashingTwin.AttackManager.Reset();
-
-            if (attackManager.AiTimer >= length)
-            {
-                attackManager.CountUp = false;
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private bool Attack_AlternatingFastDashes(int length, int dashLength, float speed, int trackTime, int dashAtTime)
-    {
-        attackManager.CountUp = true;
-        
-        if (attackManager.AiTimer == 0) SoundEngine.PlaySound(SoundID.ForceRoarPitched);
-        
-        ref var turn = ref NPC.localAI[0];
-        Twin dashingTwin = turn == 0 ? Spazmatism : Retinazer;
-        Twin notDashingTwin = turn == 0 ? Retinazer : Spazmatism;
-        
-        var target = Main.player[NPC.target];
-        var dashOptions = new DashOptions
-        {
-            DashLength = dashLength,
-            DashSpeed = speed,
-            LookOffset = MathHelper.PiOver2,
-            DashAtTime = dashAtTime,
-            MinimumDistance = 0,
-            DontReposition = true,
-            TrackTime = trackTime
-        };
-
-        if (dashingTwin.AttackManager.AiTimer == 0)
-        {
-            var dest = new Vector2();
-            if (dashingTwin is Spazmatism) dest = target.Center + new Vector2(0, 350);
-            else dest = target.Center + new Vector2(0, -350);
-            
-            dashingTwin.LookTowards(dest, 1f);
-            Teleport(dashingTwin, dest, 0);
-            dashingTwin.LookTowards(target.Center, 1f);
-
-            dashingTwin.Npc.velocity = Vector2.Zero;
-        }
-        
-        var dashState = Dash(dashingTwin, dashingTwin.AttackManager, target.Center, dashOptions);
-
-        if (dashState == DashState.Done)
-        {
-            turn = ((int) turn + 1) % 2;
-            dashingTwin.AttackManager.Reset();
-
-            if (attackManager.AiTimer >= length)
-            {
-                attackManager.CountUp = false;
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private bool Attack_SweepingLaser()
-    {
-        ref var startAngle = ref NPC.localAI[0];
-        ref var deathray = ref NPC.localAI[1];
-        var npc = Retinazer.Npc;
-
-        const int rotateTime = 30;
-        const int indicateTime = rotateTime + 60;
-        const int rayTime = indicateTime + 120;
-
-        const float spreadRadius = MathHelper.PiOver4;
-
-        Attack_Hover(Spazmatism, 10f, 0.15f);
-        
-        npc.SimpleFlyMovement(Vector2.Zero, 0.5f);
-        attackManager.CountUp = true;
-        if (attackManager.AiTimer == 0)
-        {
-            startAngle = npc.rotation;
-        }
-
-        if (attackManager.AiTimer == 0 && Main.netMode != NetmodeID.MultiplayerClient)
-        {
-            NewRetSweepIndicator(npc.Center, startAngle + MathHelper.PiOver2, indicateTime);
-        }
-
-        if (attackManager.AiTimer == indicateTime && Main.netMode != NetmodeID.MultiplayerClient)
-        {
-            var pos = npc.Center + npc.rotation.ToRotationVector2() * npc.width;
-            NewRetDeathray(pos, MathHelper.PiOver2, rayTime - indicateTime);
-        }
-        
-        // Sweep indicator down
-        else if (attackManager.AiTimer < indicateTime)
-        {
-            var progress = Utils.GetLerpValue(rotateTime, indicateTime, attackManager.AiTimer, true);
-            var ease = EasingHelper.BackOut(progress);
-            
-            var offset = spreadRadius * ease;
-            npc.rotation = startAngle + offset;
-        }
-        // Sweep laser up
-        else if (attackManager.AiTimer < rayTime)
-        {
-            var progress = 1f - Utils.GetLerpValue(indicateTime, rayTime, attackManager.AiTimer, true);
-            var ease = EasingHelper.QuadInOut(progress);
-            
-            var offset = spreadRadius * (ease - 0.5f) * 2f;
-            npc.rotation = startAngle + offset;
-        }
-        // Done
-        else
-        {
-            if (Main.netMode != NetmodeID.MultiplayerClient)
-            {
-                var dr = Utilities.FindProjectileByIdentity((int) deathray, -1);
-                dr.Kill();
-            }
-            attackManager.CountUp = false;
-            return true;
-        }
-
-        return false;
-    }
-
-    private bool Attack_Test()
-    {
-        NewRetLazer(Retinazer.Npc.Center, (Retinazer.Npc.rotation + MathHelper.PiOver2).ToRotationVector2() * 20f, MathHelper.PiOver2, 30, Retinazer.Npc.whoAmI);
-        NewSpazFireball(Spazmatism.Npc.Center, (Spazmatism.Npc.rotation + MathHelper.PiOver2).ToRotationVector2() * 5f);
-        return true;
-    }
-
-    private DashState Dash(Twin twin, AttackManager am, Vector2 dashTarget, DashOptions options)
-    {
-        var npc = twin.Npc;
-        var state = DashHelper.Dash(twin.Npc, am, dashTarget, options);
-        
-        if (am.AiTimer == 0)
-        {
-            if (Main.netMode != NetmodeID.MultiplayerClient)
-            {
-                NewDashLine(twin, npc.Center, MathHelper.PiOver2, options.DashAtTime);
-            }
-        }
-        
-        // Lots of FX
-        if (state == DashState.StartingDash)
-        {
-            DashFx(twin);
-            twin.UseAfterimages = true;
-        }
-        
-        if (state == DashState.Done)
-        {
-            twin.UseAfterimages = false;
-        }
-
-        return state;
-    }
-
-    // Dash effects
-    private void DashFx(Twin twin)
-    {
-        var npc = twin.Npc;
-        
-        // Smoke ring
-        var ring = new SmokeRingParticle(npc.Center, -npc.velocity * 0.25f, npc.rotation, Color.White, 30);
-        ring.Opacity = 0.5f;
-        ring.Scale *= 2f;
-        ring.Spawn();
-
-        // Sparks
-        for (var i = 0; i < 20; i++)
-        {
-            var dustId = twin is Spazmatism ? DustID.CursedTorch : DustID.CrimsonTorch;
-            var velOffset = Main.rand.NextVector2Circular(10f, 10f);
-                
-            Dust.NewDustDirect(npc.Center, 0, 0, dustId, 
-                -(npc.velocity.X * 0.75f) + velOffset.X,
-                -(npc.velocity.Y * 0.75f) + velOffset.Y, Scale: 1.5f);
-        }
-            
-        // Smoke dust
-        for (var i = 0; i < 20; i++)
-        {
-            var velOffset = Main.rand.NextVector2Circular(15f, 15f);
-            Dust.NewDustDirect(npc.Center, 0, 0, DustID.Smoke, 
-                -(npc.velocity.X * 0.75f) + velOffset.X,
-                -(npc.velocity.Y * 0.75f) + velOffset.Y, Scale: 2f);
-        }
-
-        // Lighting
-        var color = twin is Spazmatism ? Color.Lime : Color.Red;
-        Lighting.AddLight(npc.Center - npc.velocity * 5f, color.ToVector3());
-            
-        SoundEngine.PlaySound(SoundID.Item89, npc.Center);
-    }
-
-    private void Teleport(Twin twin, Vector2 position, float recoil)
-    {
-        var npc = twin.Npc;
-        var awayDir = npc.DirectionTo(position);
-        var startPos = npc.Center;
-
-        npc.rotation = awayDir.ToRotation() - MathHelper.PiOver2;
-        DashFx(twin);
-        
-        if (Main.netMode != NetmodeID.MultiplayerClient)
-        {
-            NewAfterimage(twin, startPos, position);
-        }
-        
-        npc.Center = position;
-        npc.velocity = awayDir * recoil;
-
-        var disperse = new BigSmokeDisperseParticle(npc.Center, Vector2.Zero, 0f, Color.WhiteSmoke, 30);
-        disperse.Scale *= 2f;
-        disperse.Opacity = 0.5f;
-        disperse.Spawn();
-    }
-    
-    #endregion
-    
-    #region Projectiles
-    
-    private Projectile NewDashLine(Twin twin, Vector2 position, float offset, int lifetime, bool anchorToBoss = true)
-    {
-        var ai1 = anchorToBoss ? twin.Npc.whoAmI : -1;
-        return BaseLineProjectile.Create<TwinsDashLine>(NPC.GetSource_FromAI(), position, offset, lifetime, ai1);
-    }
-    
-    private Projectile NewAfterimage(Twin twin, Vector2 startPos, Vector2 endPos)
-    {
-        return NpcAfterimageTrail.Create(NPC.GetSource_FromAI(), startPos, endPos, twin.Npc.whoAmI);
-    }
-
-    private Projectile NewSpazFlamethrower(Vector2 pos, float rotation)
-    {
-        return BaseBetsyFlame.Create<SpazFlamethrower>(NPC.GetSource_FromAI(), pos, rotation - MathHelper.PiOver2, Spazmatism.Npc.damage, 4, Spazmatism.Npc.whoAmI);
-    }
-
-    private Projectile NewSpazFireball(Vector2 pos, Vector2 vel)
-    {
-        return ProjHelper.NewUnscaledProjectile(NPC.GetSource_FromAI(), pos, vel, ProjectileID.CursedFlameHostile,
-            Spazmatism.Npc.damage, 3);
-    }
-
-    private Projectile NewRetDeathray(Vector2 position, float rotation, int lifetime)
-    {
-        return DeathrayBase.Create<RetDeathray>(NPC.GetSource_FromAI(), position, Retinazer.Npc.damage, 3, rotation,
-            lifetime, Retinazer.Npc.whoAmI);
-    }
-
-    private Projectile NewRetSweepIndicator(Vector2 pos, float rotation, int lifetime)
-    {
-        return BaseSweep.Create<RetSweepIndicator>(NPC.GetSource_FromAI(), pos, rotation, lifetime,
-            Retinazer.Npc.whoAmI);
-    }
-    
-    private Projectile NewRetLazer(Vector2 pos, Vector2 vel, float rotation, int lifetime, int anchor = -1)
-    {
-        return BaseLineProjectile.Create<RetLaserIndicator>(NPC.GetSource_FromAI(), pos, vel, Retinazer.Npc.damage, 3, rotation, lifetime, anchor);
-    }
-    
-    #endregion
-
-    private static void FlyTo(NPC npc, Vector2 target, float speed, float acceleration)
-    {
-        var distanceLerp = Utils.GetLerpValue(400, 200, npc.Distance(target), true);
-        var spd = MathHelper.Lerp(speed, 1f, distanceLerp);
-        var accel = MathHelper.Lerp(acceleration, acceleration * 2, distanceLerp);
-        npc.SimpleFlyMovement(npc.DirectionTo(target) * spd, accel);
-    }
 
     public override void SendAcidAI(BinaryWriter binaryWriter)
     {
